move on functionality for separate service object and updated tests
