+++ app/interactions/actions/domain_update.rb
@@ -14,7 +14,7 @@ module Actions
       assign_new_registrant if params[:registrant]
       assign_relational_modifications
       assign_requested_statuses
-      validate_dnssec unless Rails.env.test?
+      ValidateDnssec.validate_dnssec(params: params, domain: domain)
       ::Actions::BaseAction.maybe_attach_legal_doc(domain, params[:legal_document])
 
       commit
@@ -112,98 +112,6 @@ module Actions
       else
         domain.add_epp_error(2005, nil, nil, %i[dnskeys invalid])
       end
-    end
-
-    def prepare_resolver
-      dns_servers = ENV['dnssec_resolver_ips'].to_s.split(',').map(&:strip)
-      dns = Dnsruby::Resolver.new({ nameserver: dns_servers })
-      dns.do_validation = true
-      dns.do_caching = true
-      dns.dnssec = true
-
-      dns
-    end
-
-    def validate_dnssec
-      return if @params[:action] == 'rem'
-
-      dns = prepare_resolver
-      subzone_records = get_dnskey_records_from_subzone(resolver: dns, hostname: @params[:domain])
-      form_extension_records = extensional_dnskeys_data
-
-      return true if form_extension_records.empty?
-
-      validate_data(subzone_records: subzone_records, form_extension_records: form_extension_records)
-    end
-
-    def make_magic(subzone_records:, form_data:)
-      subzone_records.any? do |subzone_data|
-        subzone_data[:basic] == form_data[:basic] &&
-          subzone_data[:public_key].include?(form_data[:public_key])
-      end
-    end
-
-    def validate_data(subzone_records:, form_extension_records:)
-      flag = false
-      form_extension_records.each do |form_data|
-        flag = make_magic(subzone_records: subzone_records, form_data: form_data)
-
-        break if flag
-      end
-
-      return validation_dns_key_error unless flag
-
-      flag
-    end
-
-    def get_dnskey_records_from_subzone(resolver:, hostname:)
-      ds_records_answers = resolver.query(hostname, 'DNSKEY').answer
-
-      result_container = []
-
-      ds_records_answers.each do |ds|
-        next unless ds.type == Dnsruby::Types.DNSKEY
-
-        result_container << {
-          basic: {
-            flags: ds.flags.to_s,
-            algorithm: ds.algorithm.code.to_s,
-            protocol: ds.protocol.to_s,
-          },
-          public_key: ds.public_key.export.gsub!(/\s+/, ''),
-        }
-      end
-
-      result_container
-    rescue Dnsruby::NXDomain
-      domain.add_epp_error('2308', nil, nil, I18n.t(:dns_policy_violation))
-    end
-
-    def validation_dns_key_error
-      domain.add_epp_error('2308', nil, nil, I18n.t(:dns_policy_violation))
-    end
-
-    def extensional_dnskeys_data
-      dnskeys_data = @params[:dns_keys]
-
-      return [] if dnskeys_data.nil?
-
-      result_container = []
-
-      dnskeys_data.each do |ds|
-        next if ds[:action] == 'rem'
-
-        result_container << {
-          basic: {
-            flags: ds[:flags].to_s,
-            algorithm: ds[:alg].to_s,
-            protocol: ds[:protocol].to_s,
-          },
-          public_key: ds[:public_key],
-        }
-      end
-
-      result_container
     end
 
     def assign_removable_dnskey(key)
