+++ test/integration/epp/domain/update/base_test.rb
@@ -10,6 +10,7 @@ class EppDomainUpdateBaseTest < EppTestCase
     @original_registrant_change_verification =
     Setting.request_confirmation_on_registrant_change_enabled
     ActionMailer::Base.deliveries.clear
+    Spy.on_instance_method(ValidateDnssec, :validate_dnssec).and_return(true)
   end
 
   teardown do
