+++ test/jobs/domain_update_confirm_job_test.rb
@@ -14,6 +14,7 @@ class DomainUpdateConfirmJobTest < ActiveSupport::TestCase
                                     new_registrant_name: @new_registrant.name,
                                     new_registrant_email: @new_registrant.email,
                                     current_user_id: @user.id })
+    Spy.on_instance_method(ValidateDnssec, :validate_dnssec).and_return(true)
   end
 
   def teardown
